Add extra security to edit_booking & delete_booking views

+++ bookingsystem/views.py
@@ -1,9 +1,12 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.views import generic
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.views.generic.edit import FormView
+from django.contrib.auth.models import User
 from .forms import OnlineForm, SignUpForm
 from .models import Booking
+
 
 
 class Home(generic.DetailView):
@@ -102,6 +105,7 @@ class ListBookingView(generic.DetailView):
             return redirect('account_login')
 
 
+@login_required
 def edit_booking_view(request, booking_id):
     """
     When a user is on the My Bookings page
@@ -114,21 +118,30 @@ def edit_booking_view(request, booking_id):
     they will be redirected to the home page and a
     confimation message will appear.
     """
-    booking = get_object_or_404(Booking, id=booking_id)
-    if request.method == 'POST':
-        form = OnlineForm(data=request.POST, instance=booking)
-        if form.is_valid():
-            form.save()
-            # Pops up a message to the user when a booking is edited
-            messages.success(request, 'Your booking has been updated')
+
+    if request.user.is_authenticated:
+        user = Booking.objects.filter(user=request.user).first()
+        booking = get_object_or_404(Booking, id=booking_id)
+        if user == booking:
+            if request.method == 'POST':
+                form = OnlineForm(data=request.POST, instance=booking)
+                if form.is_valid():
+                    form.save()
+                    # Pops up a message to the user when a booking is edited
+                    messages.success(request, 'Your booking has been updated')
+                    return redirect('/')
+        else:
+            messages.error(request, 'You do not have the authority to access this page!')
             return redirect('/')
+
     form = OnlineForm(instance=booking)
 
     return render(request, 'edit_bookings.html', {
         'form': form
-    })
+        })
 
 
+@login_required
 def delete_booking(request, booking_id):
     """
     When a user is on the My Bookings page
@@ -138,11 +151,17 @@ def delete_booking(request, booking_id):
     booking id, redirect the user back to the home page and
     pop up a confimation message will appear.
     """
-    booking = get_object_or_404(Booking, id=booking_id)
-    booking.delete()
-    # Pops up a message to the user when a bookings is cancelled
-    messages.success(request, 'Your booking has been cancelled')
-    return redirect('/')
+    if request.user.is_authenticated:
+        user = Booking.objects.filter(user=request.user).first()
+        booking = get_object_or_404(Booking, id=booking_id)
+        if user == booking:
+            booking.delete()
+            # Pops up a message to the user when a bookings is cancelled
+            messages.success(request, 'Your booking has been cancelled')
+            return redirect('/')
+        else:
+            messages.error(request, 'You do not have the authority to access this page!')
+            return redirect('/')
 
 
 class SignUpView(FormView):
